feat(nexus-web): migrate account pages to semantic CSS tokens
Extract profile, api-keys, and appearance into dedicated components with vanilla CSS and shared design tokens for account layout and shell consistency.
