+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/app/account/account-layout.css
@@ -1,0 +1,30 @@
+/* account layout — sidebar nav semantic styles */
+
+.account-nav-item {
+  display: flex;
+  align-items: center;
+  gap: var(--space-3);
+  padding: var(--space-2) var(--space-3);
+  border-radius: var(--radius-md);
+  font-size: var(--font-size-sm);
+  font-weight: var(--font-weight-medium);
+  line-height: var(--line-height-sm);
+  color: var(--muted-foreground-hex);
+  text-decoration: none;
+  transition: background-color 0.15s ease, color 0.15s ease;
+}
+
+.account-nav-item:hover {
+  background-color: var(--muted-bg-hex);
+  color: var(--foreground-hex);
+}
+
+.account-nav-item-active {
+  background-color: #34D3991A;
+  color: var(--color-primary);
+}
+
+.account-nav-item-active:hover {
+  background-color: #34D3991A;
+  color: var(--color-primary);
+}
