+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/app/account/appearance/appearance.css
@@ -1,0 +1,155 @@
+/* account/appearance — semantic page styles (vanilla CSS, theme tokens) */
+
+.account-appearance-page {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-8);
+}
+
+/* Header */
+.account-appearance-header {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-1);
+}
+
+.account-appearance-title {
+  margin: 0;
+  font-size: var(--font-size-lg);
+  font-weight: var(--font-weight-semibold);
+  line-height: var(--line-height-lg);
+  color: var(--foreground-hex);
+}
+
+.account-appearance-subtitle {
+  margin: 0;
+  font-size: var(--font-size-sm);
+  line-height: var(--line-height-sm);
+  color: var(--muted-foreground-hex);
+}
+
+/* Theme section */
+.account-appearance-theme-section {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-4);
+}
+
+.account-appearance-theme-heading {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-1);
+}
+
+.account-appearance-theme-label {
+  margin: 0;
+  font-weight: var(--font-weight-medium);
+  color: var(--foreground-hex);
+}
+
+.account-appearance-theme-description {
+  margin: 0;
+  font-size: var(--font-size-sm);
+  line-height: var(--line-height-sm);
+  color: var(--muted-foreground-hex);
+}
+
+.account-appearance-theme-grid {
+  display: grid;
+  gap: var(--space-4);
+  grid-template-columns: 1fr;
+}
+
+@media (min-width: 640px) {
+  .account-appearance-theme-grid {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
+
+@media (min-width: 768px) {
+  .account-appearance-theme-grid {
+    grid-template-columns: repeat(4, 1fr);
+  }
+}
+
+.account-appearance-theme-option {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: var(--space-3);
+  padding: var(--space-4);
+  border: 1px solid var(--border-hex);
+  border-radius: var(--radius-lg);
+  background-color: var(--card-hex);
+  text-align: left;
+  cursor: pointer;
+  transition: border-color 0.15s ease, background-color 0.15s ease;
+}
+
+.account-appearance-theme-option:hover {
+  border-color: color-mix(in srgb, var(--color-primary) 50%, var(--border-hex));
+  background-color: var(--muted-bg-hex);
+}
+
+.account-appearance-theme-option--selected {
+  border-color: var(--color-primary);
+  background-color: color-mix(in srgb, var(--color-primary) 10%, transparent);
+}
+
+.account-appearance-theme-option-check {
+  position: absolute;
+  top: var(--space-3);
+  right: var(--space-3);
+  color: var(--color-primary);
+}
+
+.account-appearance-theme-option-icon {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 40px;
+  height: 40px;
+  border-radius: var(--radius-lg);
+  background-color: var(--muted-bg-hex);
+  color: var(--muted-foreground-hex);
+}
+
+.account-appearance-theme-option--selected .account-appearance-theme-option-icon {
+  background-color: color-mix(in srgb, var(--color-primary) 10%, transparent);
+  color: var(--color-primary);
+}
+
+.account-appearance-theme-option-text {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-1);
+}
+
+.account-appearance-theme-option-name {
+  margin: 0;
+  font-weight: var(--font-weight-medium);
+  color: var(--foreground-hex);
+}
+
+.account-appearance-theme-option-detail {
+  margin: 0;
+  font-size: var(--font-size-xs);
+  line-height: var(--line-height-xs);
+  color: var(--muted-foreground-hex);
+}
+
+/* Hint */
+.account-appearance-hint {
+  padding: var(--space-4);
+  border: 1px solid var(--border-hex);
+  border-radius: var(--radius-lg);
+  background-color: color-mix(in srgb, var(--muted-bg-hex) 30%, transparent);
+}
+
+.account-appearance-hint-text {
+  margin: 0;
+  font-size: var(--font-size-sm);
+  line-height: var(--line-height-sm);
+  color: var(--muted-foreground-hex);
+}
