+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/app/account/page.tsx
@@ -1,335 +1,5 @@
-'use client';
+import { redirect } from 'next/navigation';
 
-import { useState, useEffect, useRef } from 'react';
-import { Camera, Save, X, RefreshCw } from 'lucide-react';
-import { cn } from '@/lib/utils';
-import { getApiUrl } from '@/lib/config';
-import { useAuthStore } from '@/stores/auth';
-
-const BIO_MAX_LENGTH = 2000;
-
-type FastApiValidationError = {
-  loc?: (string | number)[];
-  msg?: string;
-};
-
-function formatApiError(err: unknown, fallback: string): string {
-  if (!err || typeof err !== 'object') return fallback;
-  const detail = (err as { detail?: unknown }).detail;
-  if (typeof detail === 'string') return detail;
-  if (Array.isArray(detail)) {
-    const messages = (detail as FastApiValidationError[])
-      .map((d) => {
-        const field = Array.isArray(d?.loc)
-          ? d.loc.filter((p) => p !== 'body').join('.')
-          : '';
-        return field ? `${field}: ${d?.msg ?? ''}`.trim() : (d?.msg ?? '');
-      })
-      .filter(Boolean);
-    if (messages.length) return messages.join('\n');
-  }
-  return fallback;
-}
-
-export default function ProfilePage() {
-  const authUser = useAuthStore((s) => s.user);
-  const setUser = useAuthStore((s) => s.setUser);
-  const [name, setName] = useState('');
-  const [email, setEmail] = useState('');
-  const [company, setCompany] = useState('');
-  const [role, setRole] = useState('');
-  const [bio, setBio] = useState('');
-  const [saved, setSaved] = useState(false);
-  const [uploading, setUploading] = useState(false);
-  const [avatarUrl, setAvatarUrl] = useState('');
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
-  // Initialize from authenticated user
-  useEffect(() => {
-    if (authUser) {
-      setName(authUser.name || '');
-      setEmail(authUser.email || '');
-      setCompany(authUser.company || '');
-      setRole(authUser.role || '');
-      setBio(authUser.bio || '');
-      setAvatarUrl(authUser.avatar || '');
-    }
-  }, [authUser]);
-
-  const handleAvatarUpload = async (event: React.ChangeEvent<HTMLInputElement>) => {
-    const file = event.target.files?.[0];
-    if (!file) return;
-
-    // Validate file type
-    const validTypes = ['image/png', 'image/jpeg', 'image/jpg', 'image/gif', 'image/webp'];
-    if (!validTypes.includes(file.type)) {
-      alert('Please upload a valid image file (PNG, JPG, GIF, or WEBP)');
-      return;
-    }
-
-    // Validate file size (2MB max)
-    if (file.size > 2 * 1024 * 1024) {
-      alert('File size must be less than 2MB');
-      return;
-    }
-
-    setUploading(true);
-    try {
-      const { authFetch } = await import('@/stores/auth');
-      const formData = new FormData();
-      formData.append('file', file);
-
-      const response = await authFetch(
-        `/api/auth/upload-avatar`,
-        {
-          method: 'POST',
-          body: formData,
-        }
-      );
-
-      if (response.ok) {
-        const data = await response.json();
-        const fullUrl = `${getApiUrl()}${data.avatar_url}`;
-        setAvatarUrl(fullUrl);
-        
-        // Update user in auth store
-        if (authUser) {
-          setUser({ ...authUser, avatar: fullUrl });
-        }
-      } else {
-        const error = await response.json().catch(() => ({}));
-        alert(`Upload failed: ${formatApiError(error, 'Unknown error')}`);
-      }
-    } catch (error) {
-      console.error('Upload error:', error);
-      alert('Failed to upload avatar');
-    } finally {
-      setUploading(false);
-      // Reset input
-      if (event.target) {
-        event.target.value = '';
-      }
-    }
-  };
-
-  const handleAvatarRemove = async () => {
-    if (!avatarUrl) return;
-    try {
-      const { authFetch } = await import('@/stores/auth');
-      const response = await authFetch('/api/auth/avatar', { method: 'DELETE' });
-      if (response.ok) {
-        setAvatarUrl('');
-        if (authUser) {
-          setUser({ ...authUser, avatar: undefined });
-        }
-      } else {
-        const error = await response.json().catch(() => ({}));
-        alert(`Remove failed: ${formatApiError(error, 'Unknown error')}`);
-      }
-    } catch (error) {
-      console.error('Remove avatar error:', error);
-      alert('Failed to remove avatar');
-    }
-  };
-
-  const handleSave = async () => {
-    try {
-      const { authFetch } = await import('@/stores/auth');
-      // Only include fields the user actually filled in. The backend's
-      // UserUpdate enforces min_length on name/email, so sending "" would
-      // fail validation even when the user only wanted to update bio/role.
-      const payload: Record<string, string> = {};
-      const trimmedName = name.trim();
-      const trimmedEmail = email.trim();
-      if (trimmedName) payload.name = trimmedName;
-      if (trimmedEmail) payload.email = trimmedEmail;
-      payload.company = company.trim();
-      payload.role = role.trim();
-      payload.bio = bio.trim();
-
-      const response = await authFetch('/api/auth/me', {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(payload),
-      });
-
-      if (!response.ok) {
-        const err = await response.json().catch(() => ({}));
-        alert(formatApiError(err, 'Failed to save profile'));
-        return;
-      }
-
-      const updated = await response.json();
-      const normalizeAvatar = (a?: string) =>
-        a && a.startsWith('/') ? `${getApiUrl()}${a}` : a;
-      // Update local store
-      if (authUser) {
-        setUser({
-          ...authUser,
-          name: updated.name,
-          email: updated.email,
-          avatar: normalizeAvatar(updated.avatar),
-          // Optional fields
-          // @ts-ignore - store model extended
-          company: updated.company,
-          // @ts-ignore
-          role: updated.role,
-          // @ts-ignore
-          bio: updated.bio,
-        } as any);
-      }
-      setSaved(true);
-      setTimeout(() => setSaved(false), 2000);
-    } catch (e) {
-      console.error(e);
-      alert('Failed to save profile');
-    }
-  };
-
-  return (
-    <div className="space-y-8">
-      <div>
-        <h2 className="text-lg font-semibold">Profile</h2>
-        <p className="text-sm text-muted-foreground">
-          Manage your personal information
-        </p>
-      </div>
-
-      {/* Avatar section */}
-      <div className="flex items-center gap-6">
-        <div className="relative">
-          <input
-            ref={fileInputRef}
-            type="file"
-            accept="image/png,image/jpeg,image/jpg,image/gif,image/webp"
-            onChange={handleAvatarUpload}
-            className="hidden"
-          />
-          {avatarUrl ? (
-            // eslint-disable-next-line @next/next/no-img-element
-            <img
-              src={avatarUrl}
-              alt={name}
-              className="h-24 w-24 rounded-full object-cover"
-            />
-          ) : (
-            <div className="flex h-24 w-24 items-center justify-center rounded-full bg-blue-500 text-3xl font-bold text-white">
-              {name.charAt(0).toUpperCase()}
-            </div>
-          )}
-          <button
-            onClick={() => fileInputRef.current?.click()}
-            disabled={uploading}
-            className="absolute bottom-0 right-0 flex h-8 w-8 items-center justify-center rounded-full bg-secondary text-muted-foreground hover:bg-secondary/80 disabled:opacity-50 disabled:cursor-not-allowed"
-            title="Upload avatar"
-          >
-            {uploading ? (
-              <RefreshCw size={16} className="animate-spin" />
-            ) : (
-              <Camera size={16} />
-            )}
-          </button>
-        </div>
-        <div>
-          <h3 className="font-medium">{name}</h3>
-          <p className="text-sm text-muted-foreground">{email}</p>
-          <p className="mt-1 text-xs text-muted-foreground">
-            Click camera icon to upload avatar (max 2MB)
-          </p>
-          {avatarUrl && (
-            <button
-              onClick={handleAvatarRemove}
-              className="mt-2 flex items-center gap-1 text-xs text-destructive hover:underline"
-            >
-              <X size={12} />
-              Remove picture
-            </button>
-          )}
-        </div>
-      </div>
-
-      {/* Form */}
-      <div className="space-y-6 rounded-xl border bg-card p-6">
-        <div className="grid gap-6 sm:grid-cols-2">
-          <div>
-            <label className="mb-1.5 block text-sm font-medium">Full Name</label>
-            <input
-              type="text"
-              value={name}
-              onChange={(e) => setName(e.target.value)}
-              className="w-full rounded-lg border bg-background px-3 py-2 text-sm outline-none focus:ring-2 focus:ring-blue-500/30"
-            />
-          </div>
-          <div>
-            <label className="mb-1.5 block text-sm font-medium">Email</label>
-            <input
-              type="email"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              className="w-full rounded-lg border bg-background px-3 py-2 text-sm outline-none focus:ring-2 focus:ring-blue-500/30"
-            />
-          </div>
-          <div>
-            <label className="mb-1.5 block text-sm font-medium">Company</label>
-            <input
-              type="text"
-              value={company}
-              onChange={(e) => setCompany(e.target.value)}
-              placeholder="Your company"
-              className="w-full rounded-lg border bg-background px-3 py-2 text-sm outline-none focus:ring-2 focus:ring-blue-500/30"
-            />
-          </div>
-          <div>
-            <label className="mb-1.5 block text-sm font-medium">Role</label>
-            <input
-              type="text"
-              value={role}
-              onChange={(e) => setRole(e.target.value)}
-              placeholder="Your role"
-              className="w-full rounded-lg border bg-background px-3 py-2 text-sm outline-none focus:ring-2 focus:ring-blue-500/30"
-            />
-          </div>
-        </div>
-        <div>
-          <div className="mb-1.5 flex items-center justify-between">
-            <label className="block text-sm font-medium">Bio</label>
-            <span
-              className={cn(
-                'text-xs',
-                bio.length > BIO_MAX_LENGTH
-                  ? 'text-destructive'
-                  : 'text-muted-foreground',
-              )}
-            >
-              {bio.length} / {BIO_MAX_LENGTH}
-            </span>
-          </div>
-          <textarea
-            value={bio}
-            onChange={(e) => setBio(e.target.value.slice(0, BIO_MAX_LENGTH))}
-            placeholder="Tell us about yourself..."
-            rows={12}
-            maxLength={BIO_MAX_LENGTH}
-            className="w-full resize-y rounded-lg border bg-background px-3 py-2 text-sm leading-relaxed outline-none focus:ring-2 focus:ring-blue-500/30"
-          />
-        </div>
-      </div>
-
-      {/* Save button */}
-      <div className="flex justify-end">
-        <button
-          onClick={handleSave}
-          className={cn(
-            'flex items-center gap-2 rounded-lg px-4 py-2 text-sm font-medium transition-colors',
-            saved
-              ? 'bg-blue-500/20 text-blue-500'
-              : 'bg-blue-500 text-white hover:bg-blue-600'
-          )}
-        >
-          <Save size={16} />
-          {saved ? 'Saved!' : 'Save Changes'}
-        </button>
-      </div>
-    </div>
-  );
+export default function AccountPage() {
+  redirect('/account/profile');
 }
