+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/components/shell/header.tsx
@@ -241,7 +241,7 @@ export function Header({ title, subtitle, actions }: HeaderProps = {}) {
               {/* Menu items */}
               <div className="py-2">
                 <Link
-                  href="/account"
+                  href="/account/profile"
                   onClick={() => setUserMenuOpen(false)}
                   className="flex items-center gap-3 rounded-md px-4 py-2.5 text-sm transition-colors hover:bg-muted"
                 >
