+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/components/shell/sidebar/settings-section.tsx
@@ -8,6 +8,7 @@ import { useWorkspaceStore } from '@/stores/workspace';
 import { CollapsibleSection } from './collapsible-section';
 import { getWorkspacePath } from './utils';
 import { SETTINGS_GROUPS } from '@/components/shell/settings-nav';
+import { shellTokens } from '../tokens';
 
 export function SettingsSection({ collapsed, detailOnly }: { collapsed: boolean; detailOnly?: boolean }) {
   const pathname = usePathname();
@@ -39,7 +40,8 @@ export function SettingsSection({ collapsed, detailOnly }: { collapsed: boolean;
                   key={item.href}
                   href={fullHref}
                   className={cn(
-                    'flex w-full items-center gap-2 rounded-md px-2 py-1.5 text-sm leading-snug transition-colors',
+                    'flex w-full items-center gap-2 rounded-md px-2 py-1.5 transition-colors',
+                    shellTokens.sidebar.listRow,
                     isActive
                       ? 'bg-workspace-accent-15 text-workspace-accent'
                       : 'text-muted-foreground hover:bg-workspace-accent-10 hover:text-foreground'
